+++ tests/trio100.py
@@ -33,3 +33,21 @@ async def function_name():
     async with trio.fail_after(10):
         async for _ in receive_channel:
             pass
+
+
+import trio
+
+
+async def function_name():
+    with (
+        open("") as _,
+        trio.fail_after(10),  # error: 8, "trio.fail_after"
+    ):
+        pass
+
+    with (
+        trio.fail_after(5),  # error: 8, "trio.fail_after"
+        open("") as _,
+        trio.move_on_after(5),  # error: 8, "trio.move_on_after"
+    ):
+        pass